fixed routes. set user rights with Gates(). set a many to many relationship between users and locations. changed /home routing. implemented guard and company edit.

+++ routes/web.php
@@ -13,42 +13,44 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
+Auth::routes();
 
 Route::get('/', function () {
     return redirect('/home');
 });
 
-Auth::routes();
-
 Route::get('/home', 'HomeController@index');
 
-Route::get('/profile/{user}', 'ProfilesController@index')->name('profile');
+Route::get('/profile/{user}', 'ProfilesController@index')->middleware('can:use-application')->name('profile');
 
 Route::get('/shift/index', 'ShiftsController@index')->middleware('can:view-shifts')->name('shift.index');
 Route::get('/shift/create', 'ShiftsController@create')->middleware('can:create-shifts');
 Route::get('/shift/{shift}/edit', 'ShiftsController@edit')->middleware('can:edit-shifts')->name('shift.edit');
-Route::get('/shift/{shift}', 'ShiftsController@show')->name('shift.show');
-Route::delete('/shift/{shift}', 'ShiftsController@destroy')->name('shift.destroy');
-Route::patch('/shift/{shift}', 'ShiftsController@update')->name('shift.update');
-Route::post('/sh', 'ShiftsController@store');
+Route::delete('/shift/{shift}', 'ShiftsController@destroy')->middleware('can:admin')->name('shift.destroy');
+Route::patch('/shift/{shift}', 'ShiftsController@update')->middleware('can:edit-shifts')->name('shift.update');
+Route::post('/sh', 'ShiftsController@store')->middleware('can:create-shifts');
 
-Route::get('/active-shift/index', 'ActiveShiftsController@index')->name('active-shift.index');
-Route::get('/active-shift/create/{shift}', 'ActiveShiftsController@create')->name('active-shift.create');
-Route::post('/as', 'ActiveShiftsController@store');
-Route::patch('/active-shift/{shift}/confirm', 'ActiveShiftsController@confirmActiveShift');
-Route::delete('/active-shift/{activeShift}', 'ActiveShiftsController@destroy')->name('active-shift.destroy');
+Route::get('/active-shift/index', 'ActiveShiftsController@index')->middleware('can:view-shifts')->name('active-shift.index');
+Route::get('/active-shift/create/{shift}', 'ActiveShiftsController@create')->middleware('can:edit-shifts')->name('active-shift.create');
+Route::post('/as', 'ActiveShiftsController@store')->middleware('can:edit-shifts');
+Route::patch('/active-shift/{shift}/confirm', 'ActiveShiftsController@confirmActiveShift')->middleware('can:confirm-shifts');
+Route::delete('/active-shift/{activeShift}', 'ActiveShiftsController@destroy')->middleware('can:admin')->name('active-shift.destroy');
 
-Route::get('/security', 'SecurityController@index')->middleware('can:manage-security')->name('company.index');
+Route::get('/security/index', 'SecurityController@index')->middleware('can:manage-security')->name('company.index');
 Route::get('/security/create', 'SecurityController@create')->middleware('can:manage-security');
 Route::delete('/security/{company}', 'SecurityController@destroy')->middleware('can:admin')->name('company.destroy');
 Route::get('/security/{company}/edit', 'SecurityController@edit')->middleware('can:manage-security')->name('company.edit');
 Route::post('/s', 'SecurityController@store')->middleware('can:manage-security');
+Route::patch('/s/{company}', 'SecurityController@update')->middleware('can:manage-security');
 
-Route::get('/guard/{company}/create', 'GuardsController@create');
-Route::delete('/guard/{guard}', 'GuardsController@destroy')->name('guard.destroy');
-Route::get('/guard/{guard}', 'GuardsController@show')->name('guard.show');
-Route::post('/g', 'GuardsController@store');
-Route::post('/guard/{guard}', 'GuardsController@exportCsv')->name('guard.exportCsv');
+Route::get('/guard/{company}/create', 'GuardsController@create')->middleware('can:create-guard');
+Route::get('/guard/{company}/index', 'GuardsController@index')->middleware('can:manage-security')->name('guard.index');
+Route::get('/guard/{guard}/edit', 'GuardsController@edit')->middleware('can:manage-security')->name('guard.edit');
+Route::delete('/guard/{guard}', 'GuardsController@destroy')->middleware('can:admin')->name('guard.destroy');
+Route::get('/guard/{guard}', 'GuardsController@show')->middleware('can:manage-shifts')->name('guard.show');
+Route::post('/g', 'GuardsController@store')->middleware('can:create-guard');
+Route::patch('/g/{guard}', 'GuardsController@update')->middleware('can:create-guard');
+Route::post('/guard/{guard}', 'GuardsController@exportCsv')->middleware('can:manage-shifts')->name('guard.exportCsv');
 
 Route::namespace('Admin')
     ->prefix('admin')
